Reorganised file sturcture
