--- Properties/launchSettings.json
@@ -1,11 +1,0 @@
-{
-  "profiles": {
-    "bordertale": {
-      "commandName": "Project"
-    },
-    "WSL 2": {
-      "commandName": "WSL2",
-      "distributionName": ""
-    }
-  }
-}
