+++ bordertale.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bordertale", "bordertale.csproj", "{A9141147-9668-4370-805F-64351C86CC16}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bordertale", ".\bordertale\bordertale.csproj", "{A9141147-9668-4370-805F-64351C86CC16}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
