+++ bordertale/Properties/launchSettings.json
@@ -1,0 +1,11 @@
+{
+  "profiles": {
+    "bordertale": {
+      "commandName": "Project"
+    },
+    "WSL 2": {
+      "commandName": "WSL2",
+      "distributionName": ""
+    }
+  }
+}
