+++ bordertale/bordertale.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <ApplicationIcon>media\icon.ico</ApplicationIcon>
+  </PropertyGroup>
+
+</Project>
